Fixed attribute modifiers

+++ common/src/main/java/banduty/stoneycore/util/definitionsloader/ArmorDefinitionData.java
@@ -8,12 +8,12 @@ import net.minecraft.network.codec.StreamCodec;
 
 import java.util.Map;
 
-public record ArmorDefinitionData(Map<String, Double> damageResistance, float deflectChance, double weight) {
+public record ArmorDefinitionData(Map<String, Double> damageResistance, double deflectChance, double weight) {
     public static final Codec<ArmorDefinitionData> CODEC = RecordCodecBuilder.create(instance -> instance.group(
             Codec.unboundedMap(Codec.STRING.xmap(String::toUpperCase, s -> s), Codec.DOUBLE)
                     .optionalFieldOf("damageResistance", Map.of())
                     .forGetter(ArmorDefinitionData::damageResistance),
-            Codec.FLOAT.optionalFieldOf("deflectChance", 0.0f)
+            Codec.DOUBLE.optionalFieldOf("deflectChance", 0.0d)
                     .forGetter(ArmorDefinitionData::deflectChance),
             Codec.DOUBLE.optionalFieldOf("weight", 0.0)
                     .forGetter(ArmorDefinitionData::weight)
